feat: gen-json-schema
