+++ .gitignore
@@ -14,3 +14,5 @@ doc_images
 **/public/datasets
 app-dist
 .vercel
+chartinfo.json
+stoinfo_v2.json
