+++ packages/graphic-walker/package.json
@@ -4,7 +4,8 @@
   "scripts": {
     "dev:front_end": "vite --host",
     "dev": "npm run dev:front_end",
-    "build": "vite build && tsc --build",
+    "gen-schema": "node scripts/create-json-schema.js",
+    "build": "vite build && tsc --build && yarn gen-schema",
     "build:app": "vite build --config vite-app.config.ts",
     "serve": "vite preview",
     "type": "tsc src/lib.ts --declaration --emitDeclarationOnly --jsx react --esModuleInterop --outDir dist",
@@ -133,6 +134,7 @@
     "@vercel/analytics": "^0.1.8",
     "@vitejs/plugin-react": "^3.1.0",
     "styled-components": "^5.3.6",
+    "ts-json-schema-generator": "^1.5.0",
     "ts-xor": "^1.3.0",
     "typescript": "^5.2.2",
     "vite": "^4.1.1"
