--- packages/graphic-walker/public/stoinfo_v2.json
@@ -1,103 +1,0 @@
-{
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "properties": {
-        "datasets": {
-            "items": {
-                "properties": {
-                    "data": {
-                        "items": {
-                            "additionalProperties": {},
-                            "type": "object"
-                        },
-                        "type": "array"
-                    },
-                    "id": {
-                        "type": "string"
-                    },
-                    "metaId": {
-                        "type": "string"
-                    },
-                    "name": {
-                        "type": "string"
-                    }
-                },
-                "type": "object",
-                "required": [
-                    "id",
-                    "metaId",
-                    "name",
-                    "data"
-                ]
-            },
-            "type": "array"
-        },
-        "metaDict": {
-            "additionalProperties": {
-                "items": {
-                    "properties": {
-                        "analyticType": {
-                            "enum": [
-                                "dimension",
-                                "measure"
-                            ],
-                            "type": "string"
-                        },
-                        "basename": {
-                            "type": "string"
-                        },
-                        "disable": {
-                            "type": "boolean"
-                        },
-                        "fid": {
-                            "type": "string"
-                        },
-                        "key": {
-                            "type": "string"
-                        },
-                        "name": {
-                            "type": "string"
-                        },
-                        "path": {
-                            "items": {
-                                "type": "string"
-                            },
-                            "type": "array"
-                        },
-                        "semanticType": {
-                            "enum": [
-                                "nominal",
-                                "ordinal",
-                                "quantitative",
-                                "temporal"
-                            ],
-                            "type": "string"
-                        }
-                    },
-                    "type": "object",
-                    "required": [
-                        "fid",
-                        "semanticType",
-                        "analyticType"
-                    ]
-                },
-                "type": "array"
-            },
-            "type": "object"
-        },
-        "specDict": {
-            "additionalProperties": {
-                "items": {
-                    "type": "string"
-                },
-                "type": "array"
-            },
-            "type": "object"
-        }
-    },
-    "type": "object",
-    "required": [
-        "metaDict",
-        "datasets",
-        "specDict"
-    ]
-}
